Updated with Download/Stream logic Script 1.0.0.17 Plugin 1.0.0.19

+++ script.video.thelivebox/sfile.py
@@ -31,7 +31,8 @@ def exists(filename):
 
 def isfile(filename):
     if not exists(filename):
-        raise Exception('sfile.isfile error %s does not exists' % filename)
+        #raise Exception('sfile.isfile error %s does not exists' % filename)
+        return False
 
     import stat
     return stat.S_ISREG(xbmcvfs.Stat(filename).st_mode())
@@ -39,7 +40,8 @@ def isfile(filename):
 
 def isdir(folder):
     if not exists(folder):
-        raise Exception('sfile.isdir error %s does not exists' % folder)
+        #raise Exception('sfile.isdir error %s does not exists' % folder)
+        return False
 
     import stat
     return stat.S_ISDIR(xbmcvfs.Stat(folder).st_mode())
@@ -108,10 +110,13 @@ def makedirs(path):
 
 
 def delete(filename):
-    return xbmcvfs.delete(filename)
+    return remove(filename)
 
 
 def remove(filename):
+    if isdir(filename):
+        return rmtree(filename)
+
     return xbmcvfs.delete(filename)
 
 
@@ -147,15 +152,24 @@ def copy(src, dst):
 
 
 def rename(src, dst):
+    if src == dst:
+        return
+
     if not exists(src):
         return
 
     if isdir(src):
+        if src.lower() == dst.lower():
+            newSrc = src +'sfile_temp_name'
+            rename(src, newSrc)
+            src = newSrc
+        
         copytree(src, dst)
         rmtree(src)
         return
 
     return xbmcvfs.rename(src, dst)
+
 
 
 def mtime(filename):
@@ -172,6 +186,46 @@ def ctime(filename):
 
     status = xbmcvfs.Stat(filename)
     return status.st_ctime()
+
+
+def getfolder(path):
+    import os
+    path = path.replace('/', os.sep)
+    if path.endswith(os.sep):
+        path += 'filename'
+
+    try:    return path.rsplit(os.sep, 1)[0]       
+    except: return ''
+
+
+def getfilename(path):
+    import os
+    path = path.replace('/', os.sep)
+    try:    return path.rsplit(os.sep, 1)[-1]
+    except: return ''
+
+
+def removeextension(path):
+    try:    return path.rsplit('.', 1)[0]
+    except: path
+
+
+def getextension(path):
+    try:    return path.rsplit('.')[-1]
+    except: return ''
+
+
+def isempty(folder):
+    current, dirs, files = walk(folder)
+
+    if len(dirs) > 0:
+        return False
+
+    if len(files) > 0:
+        return False
+
+    return True
+
